+++ src/main/java/insane96mcp/iguanatweaksreborn/module/hungerhealth/fooddrinks/FoodDrinks.java
@@ -110,7 +110,7 @@ public class FoodDrinks extends JsonFeature {
 	public static Integer rawFoodPoisonAmplifier = 1;
 	@Config(min = 0d)
 	@Label(name = "Raw food Poison Duration Multiplier", description = "Raw food's poison duration will be multiplied by this value. With this set to 1, raw food will give 1 second of poison per nutrition + saturation given.")
-	public static Double rawFoodPoisonDurationMultiplier = 5d;
+	public static Double rawFoodPoisonDurationMultiplier = 2d;
 	@Config
 	@Label(name = "Combat Snapshot eating saturation", description = "If enabled, when eating food the saturation will not sum, instead will just be set to the food's saturation (if higher than the current)")
 	public static Boolean combatSnapshotEatingSaturation = true;